requirements.txt
# CrewAI & AI Agents
crewai>=0.28.0
crewai[tools]>=0.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Data Profiling & EDA
ydata-profiling>=4.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Application
streamlit>=1.28.0
flask>=3.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
